backend/pyproject.toml
[tool.poetry]
name = "todo-app-backend"
version = "0.1.0"
description = "Backend for the Todo App with Authentication"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlmodel = "^0.0.11"
better-exceptions = "^0.3.2"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.7.1"
black = "^23.12.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = ["sqlmodel.*", "better_auth.*"]
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--tb=short",
    "-v"
]